+++ src/components/PostTag/PostTagStyled.ts
@@ -1,0 +1,14 @@
+"use client";
+
+import styled from "styled-components";
+
+export const TagName = styled.li`
+  float: left;
+
+  font-size: 18px;
+  color: var(--gray-500);
+
+  @media (max-width: 768px) {
+    font-size: 12px;
+  }
+`;
