+++ src/components/RunnerPost/RunnerPostItem/RunnerPostItem.tsx
@@ -1,10 +1,10 @@
 import React from "react";
-import { styled } from "styled-components";
 import Avatar from "@/components/common/Avatar/Avatar";
 import { RunnerPost } from "@/types/runnerPost";
 import Label from "@/components/common/Label/Label";
 import { REVIEW_STATUS_LABEL_TEXT } from "@/constants";
 import useViewport from "@/hooks/useViewport";
+import * as S from "./RunnerPostItemStyled";
 
 const RunnerPostItem = ({
   runnerPostData: {
@@ -20,8 +20,6 @@ const RunnerPostItem = ({
 }: {
   runnerPostData: RunnerPost;
 }) => {
-  const { isMobile } = useViewport();
-
   return (
     <S.RunnerPostItemContainer>
       <S.LeftSideContainer>
@@ -29,7 +27,6 @@ const RunnerPostItem = ({
         <S.DeadLineContainer>
           <S.DeadLine>{deadline.replace("T", " ")} 까지</S.DeadLine>
           <Label
-            height={isMobile ? "18px" : "22px"}
             colorTheme={
               reviewStatus === "NOT_STARTED"
                 ? "WHITE"
@@ -37,7 +34,6 @@ const RunnerPostItem = ({
                 ? "RED"
                 : "GRAY"
             }
-            fontSize={isMobile ? "10px" : ""}
           >
             {REVIEW_STATUS_LABEL_TEXT[reviewStatus]}
           </Label>
@@ -51,11 +47,7 @@ const RunnerPostItem = ({
       <S.RightSideContainer>
         {runnerProfile ? (
           <S.ProfileContainer>
-            <Avatar
-              width={isMobile ? "30px" : "50px"}
-              height={isMobile ? "30px" : "50px"}
-              imageUrl={runnerProfile.imageUrl}
-            />
+            <Avatar imageUrl={runnerProfile.imageUrl} />
             <S.ProfileName>{runnerProfile.name}</S.ProfileName>
           </S.ProfileContainer>
         ) : null}
@@ -73,149 +65,3 @@ const RunnerPostItem = ({
 };
 
 export default RunnerPostItem;
-
-const S = {
-  RunnerPostItemContainer: styled.li`
-    display: flex;
-    justify-content: space-between;
-
-    min-width: 340px;
-    width: 100%;
-    height: max-content;
-    padding: 35px 40px;
-
-    border: 0.5px solid var(--gray-500);
-    border-radius: 12px;
-    box-shadow: 1px 2px 3px rgba(0, 0, 0, 0.2);
-
-    cursor: pointer;
-
-    &:hover {
-      transition: all 0.3s ease;
-      transform: scale(1.015);
-      outline: 1.5px solid var(--baton-red);
-    }
-
-    @media (max-width: 768px) {
-      padding: 25px 30px;
-    }
-  `,
-
-  PostTitle: styled.p`
-    margin-bottom: 15px;
-
-    font-size: 28px;
-    font-weight: 700;
-
-    @media (max-width: 768px) {
-      font-size: 16px;
-    }
-  `,
-
-  DeadLineContainer: styled.div`
-    display: flex;
-    align-items: baseline;
-    gap: 10px;
-  `,
-
-  DeadLine: styled.p`
-    margin-bottom: 60px;
-
-    color: var(--gray-600);
-
-    @media (max-width: 768px) {
-      margin-bottom: 40px;
-
-      font-size: 12px;
-    }
-  `,
-
-  TagContainer: styled.div`
-    @media (max-width: 768px) {
-      display: flex;
-      flex-wrap: wrap;
-    }
-  `,
-
-  Tag: styled.span`
-    margin-right: 10px;
-
-    font-size: 14px;
-    color: var(--gray-600);
-  `,
-
-  LeftSideContainer: styled.div``,
-
-  RightSideContainer: styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: end;
-    justify-content: space-between;
-  `,
-
-  ProfileContainer: styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: end;
-
-    margin-bottom: 30px;
-    gap: 10px;
-  `,
-
-  ProfileName: styled.p`
-    min-width: 50px;
-
-    font-size: 14px;
-    text-align: center;
-
-    @media (max-width: 768px) {
-      min-width: 30px;
-
-      font-size: 12px;
-    }
-  `,
-
-  ChatViewContainer: styled.div`
-    display: flex;
-
-    gap: 10px;
-
-    font-size: 12px;
-  `,
-
-  statisticsContainer: styled.div`
-    display: flex;
-    align-items: center;
-    gap: 5px;
-
-    margin-bottom: auto;
-
-    & > p {
-      color: #a4a4a4;
-    }
-
-    @media (max-width: 768px) {
-      gap: 2px;
-    }
-  `,
-
-  statisticsImage: styled.img`
-    width: 21px;
-    height: 20px;
-
-    margin-left: 8px;
-
-    @media (max-width: 768px) {
-      width: 16px;
-      height: 15px;
-    }
-  `,
-
-  statisticsText: styled.p`
-    font-size: 14px;
-
-    @media (max-width: 768px) {
-      font-size: 10px;
-    }
-  `,
-};
