+++ src/components/RunnerPost/RunnerPostList/RunnerPostList.tsx
@@ -1,14 +1,16 @@
 import React from "react";
-import { styled } from "styled-components";
 import RunnerPostItem from "../RunnerPostItem/RunnerPostItem";
-import { getRunnerPost } from "@/apis/apis";
+import { RunnerPost } from "@/types/runnerPost";
+import * as S from "./RunnerPostListStyled";
 
-const RunnerPostList = async () => {
-  const posts = await getRunnerPost(10);
+interface Props {
+  posts: RunnerPost[];
+}
 
+const RunnerPostList = ({ posts }: Props) => {
   return (
     <S.RunnerPostWrapper aria-label="게시글 목록">
-      {posts.data.map((runnerPostData) => (
+      {posts.map((runnerPostData) => (
         <RunnerPostItem key={runnerPostData.runnerPostId} runnerPostData={runnerPostData} />
       ))}
     </S.RunnerPostWrapper>
@@ -16,18 +18,3 @@ const RunnerPostList = async () => {
 };
 
 export default RunnerPostList;
-
-const S = {
-  RunnerPostWrapper: styled.ul`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 30px;
-
-    width: 100%;
-
-    @media (max-width: 768px) {
-      gap: 20px;
-    }
-  `,
-};
