+++ src/components/common/Tag/Tag.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import styled from 'styled-components';
-import closeIcon from '@/assets/close-icon-red.svg';
+import React from "react";
+import closeIcon from "@/assets/close-icon-red.svg";
+import * as S from "./TagStyled";
 
 interface Props extends React.HTMLProps<HTMLButtonElement> {
   children: React.ReactNode;
@@ -18,32 +18,3 @@ const Tag = ({ children, onClick }: Props) => {
 };
 
 export default Tag;
-
-const S = {
-  TagWrapper: styled.div`
-    :hover {
-      transition: all 0.35s ease;
-      transform: scale(1.04);
-    }
-  `,
-
-  Tag: styled.button`
-    display: flex;
-    align-items: center;
-
-    height: 34px;
-    padding: 0 15px;
-    gap: 5px;
-
-    background-color: white;
-    border-radius: 16px;
-    border: 1px solid var(--baton-red);
-
-    color: var(--baton-red);
-  `,
-
-  CloseTagIcon: styled.img`
-    width: 9px;
-    height: 9px;
-  `,
-};
