+++ src/layout/Layout.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import Header from './Header';
-import { styled } from 'styled-components';
+import React from "react";
+import Header from "./Header";
+import * as S from "./LayoutStyled";
 
 interface Props {
   children: React.ReactNode;
@@ -17,20 +17,3 @@ const Layout = ({ children, maxWidth }: Props) => {
 };
 
 export default Layout;
-
-const S = {
-  LayoutContainer: styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  `,
-
-  ChildrenWrapper: styled.article<{ $maxWidth?: string }>`
-    max-width: ${({ $maxWidth }) => $maxWidth || '1200px'};
-    width: 100%;
-
-    @media (max-width: 768px) {
-      padding: 15px;
-    }
-  `,
-};
